+++ src/index.ts
@@ -1,9 +1,9 @@
 export {
-  defineCacheStore,
-  type CacheStore,
-  type CacheStoreFactory,
-  type GenericCacheStore,
-  type GenericCacheStoreFactory,
-} from './defineCacheStore'
-export { defineRecordStore, watchRecordStore } from './defineRecordStore'
-export { type Options } from './storeOptions'
+  defineRecordStore,
+  type GenericRecordStore,
+  type RecordStore,
+} from './defineRecordStore'
+
+export {
+  watchRecordStore,
+} from './watchRecordStore'
