+++ tests/defineCacheStore.types.test.ts
@@ -1,13 +1,6 @@
 import { describe, expect, expectTypeOf, it } from 'vitest'
-import {
-  type CacheStore,
-  type CacheStoreFactory,
-  defineCacheStore,
-  type GenericCacheStore,
-  type GenericCacheStoreFactory,
-} from '../src'
 import type { Reactive, ToRefs } from 'vue'
-import type { Person } from './helpers/people'
+import { defineRecordStore, type GenericRecordStore, type RecordStore } from '../src'
 
 type Item = {
   id: number,
@@ -22,51 +15,17 @@ describe('defineCacheStore() types', async () => {
       getRefs(id: any): ToRefs<Reactive<Item>>,
       has(id: any): boolean,
       remove(id: any): void,
-      getUseCount(): number,
       clear(): void,
-      mount(): void,
-      unMount(): void,
     }
 
-    const useTestCache = defineCacheStore((id: number, context: CacheStore<Item>): Item => {
+    const cache: CustomCacheStore = defineRecordStore((id: number, context: RecordStore<Item>): Item => {
       return {
         id,
         name: 'susan',
       }
-    }, { autoMountAndUnMount: false })
-
-    const cache: CustomCacheStore = useTestCache()
+    })
 
     expectTypeOf(cache).toEqualTypeOf<CustomCacheStore>()
-  })
-
-
-  it('type CacheStoreFactory', async () => {
-
-    function creatorFunction(id: number): Person {
-      return {
-        id,
-        name: 'susan',
-      }
-    }
-
-    const useTestCache: CacheStoreFactory<typeof creatorFunction> = defineCacheStore(creatorFunction, { autoMountAndUnMount: false })
-    expectTypeOf(useTestCache).toEqualTypeOf<CacheStoreFactory<typeof creatorFunction>>()
-
-    const cache = useTestCache()
-    expectTypeOf(cache.get(99)).toEqualTypeOf<Person>()
-  })
-
-  it('type GenericCacheStoreFactory', async () => {
-    function creatorFunction(id: number) {
-      return {
-        id,
-        name: 'susan',
-      }
-    }
-
-    const useTestCache: GenericCacheStoreFactory = defineCacheStore(creatorFunction, { autoMountAndUnMount: false })
-    expectTypeOf(useTestCache).toEqualTypeOf<GenericCacheStoreFactory>()
   })
 
   it('type GenericCacheStore', async () => {
@@ -77,28 +36,25 @@ describe('defineCacheStore() types', async () => {
       }
     }
 
-    const useTestCache = defineCacheStore(creatorFunction, { autoMountAndUnMount: false })
-    const store: GenericCacheStore = useTestCache()
-    expectTypeOf(store).toEqualTypeOf<GenericCacheStore>()
+    const store: GenericRecordStore = defineRecordStore(creatorFunction)
+    expectTypeOf(store).toEqualTypeOf<GenericRecordStore>()
   })
 
   it('type CacheStore', () => {
     let called = 0
 
-    function creatorFunction(id: number, context: CacheStore<Item>) {
+    function creatorFunction(id: number, context: RecordStore<Item>) {
       called++
-      expectTypeOf(context).toEqualTypeOf<CacheStore<Item>>()
+      expectTypeOf(context).toEqualTypeOf<RecordStore<Item>>()
       return {
         id,
         name: 'susan',
       }
     }
 
-    const useTestCache = defineCacheStore(creatorFunction, { autoMountAndUnMount: false })
+    const cache = defineRecordStore(creatorFunction)
 
-    const cache = useTestCache()
-
-    expectTypeOf(cache).toEqualTypeOf<CacheStore<Item>>()
+    expectTypeOf(cache).toEqualTypeOf<RecordStore<Item>>()
 
     cache.get('asd')
     expect(called).toEqual(1)
