+++ tests/vitest-setup.ts
@@ -1,15 +1,4 @@
-import { afterEach, beforeEach } from 'vitest'
+import { afterEach } from 'vitest'
 import { enableAutoUnmount } from '@vue/test-utils'
-import { defineCacheStore, defineRecordStore } from '../src'
 
 enableAutoUnmount(afterEach)
-
-beforeEach(() => {
-  defineRecordStore.resetGlobalDefaultOptions()
-  defineCacheStore.resetGlobalDefaultOptions()
-})
-
-afterEach(() => {
-  defineRecordStore.resetGlobalDefaultOptions()
-  defineCacheStore.resetGlobalDefaultOptions()
-})
